+++ web-gui/src/app/modules/student/student-list/course-table/course-table-datasource.ts
@@ -1,6 +1,7 @@
-import {DataSource} from '@angular/cdk/collections';
-import {Observable} from 'rxjs';
-import {DatabaseService} from "../../../service/database.service";
+import {CollectionViewer, DataSource} from '@angular/cdk/collections';
+import {BehaviorSubject, Observable} from 'rxjs';
+import {DatabaseService} from "../../../../service/database.service";
+
 
 export interface CourseTableItem {
   course_description: string;
@@ -18,6 +19,8 @@ export interface CourseTableItem {
  */
 export class CourseTableDataSource extends DataSource<CourseTableItem> {
 
+  private courseSubject = new BehaviorSubject<CourseTableItem[]>([]);
+
   constructor(private db: DatabaseService) {
     super();
   }
@@ -27,14 +30,28 @@ export class CourseTableDataSource extends DataSource<CourseTableItem> {
    * the returned stream emits new items.
    * @returns A stream of the items to be rendered.
    */
-  connect(): Observable<CourseTableItem[]> {
-    return this.db.getCourses();
+  connect(collectionViewer: CollectionViewer): Observable<CourseTableItem[] | ReadonlyArray<CourseTableItem>> {
+    return this.courseSubject.asObservable();
   }
+
 
   /**
    *  Called when the table is being destroyed. Use this function, to clean up
    * any open connections or free any held resources that were set up during connect.
    */
-  disconnect() {
+  disconnect(collectionViewer: CollectionViewer): void {
+    this.courseSubject.complete();
   }
+
+
+  /**
+   * Load courses and emit new values.
+   */
+  loadCourses() {
+    this.db.getCourses().subscribe(value => {
+      console.log(value);
+      this.courseSubject.next(value);
+    });
+  }
+
 }
