+++ web-gui/src/app/modules/student/student-list/course-table/course-table.component.html
@@ -1,0 +1,11 @@
+<table mat-table class="full-width-table" [dataSource]="dataSource" aria-label="Elements">
+
+  <!-- Name Column -->
+  <ng-container matColumnDef="name">
+    <th mat-header-cell *matHeaderCellDef>Name</th>
+    <td mat-cell *matCellDef="let row">{{row.course_name}}</td>
+  </ng-container>
+
+  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+  <tr mat-row *matRowDef="let row; columns: displayedColumns;" (click)="getRow(row)"></tr>
+</table>
