+++ web-gui/src/app/modules/student/student-list/course-table/course-table.component.scss
