+++ web-gui/src/app/modules/student/student.module.ts
@@ -5,10 +5,11 @@ import {StudentRoutingModule} from './student-routing.module';
 import {StudentStartComponent} from "./student-start/student-start.component";
 import {MaterialComponentsModule} from "../material-components/material-components.module";
 import {MatSnackBar} from "@angular/material";
-import {CourseTableComponent} from './course-table/course-table.component';
-import { StudentSearchComponent } from './student-search/student-search.component';
-import { StudentCourseComponent } from './student-course/student-course.component';
-import { StudentListComponent } from './student-list/student-list.component';
+import {CourseTableComponent} from './student-list/course-table/course-table.component';
+import {StudentSearchComponent} from './student-search/student-search.component';
+import {StudentCourseComponent} from './student-course/student-course.component';
+import {StudentListComponent} from './student-list/student-list.component';
+import {StudentCourseDialogComponent} from "./student-course/student-course-dialog/student-course-dialog.component";
 
 /**
  * Module for everything student related.
@@ -20,7 +21,15 @@ import { StudentListComponent } from './student-list/student-list.component';
     StudentRoutingModule,
     MaterialComponentsModule,
   ],
-  declarations: [StudentStartComponent, CourseTableComponent, StudentSearchComponent, StudentCourseComponent, StudentListComponent],
+  declarations: [
+    StudentStartComponent,
+    CourseTableComponent,
+    StudentSearchComponent,
+    StudentCourseComponent,
+    StudentListComponent,
+    StudentCourseDialogComponent
+  ],
+  entryComponents: [StudentCourseDialogComponent],
   providers: [MatSnackBar]
 })
 export class StudentModule {
